emacs - stop logging into drawer
I thought this was a good idea, but usually the only content in my tasks is logging anyway.
It doesn't make much sense to put it into a drawer.

If there are tasks I want to put logging into a drawer later, I can use the LOG_INTO_DRAWER property