nvim - remove coc-angular, add coc-lists
coc-angular wasn't working for me - it's supposed to add completions and go to definition for angular templates (including external
templates) but I can't get either of those working. So it's kind of useless.

coc-lists adds support for finding files, so I don't have to use :find and guess the filename any more