projects - fix work script default project
It was broken when there was no project set, it wasn't going to the default project properly.
Could not find pj, so I need to source that file in the script.

As a note, this does not update the shell outside of TMux, it will still not have a project set.